Сделал механическую задачу полностью шаблонной, подготовил шаблонную базу для перехода на умные указатели.
